[demos] deleted old demo and rename
